- Made the warm skills BF_WEAPON type so that they may trigger effect cards. - Also changed their pl to -1 so they carry the elemeso they carry the element. - Warm skills now damage other players for 60sp per hit. - Warm skill damage interval decreased to 100ms instead of 1000ms. - TK_DODGE  now only dodges ranged weapon attacks, while under Spurt mode it dodges everything. - Cleaned up and expanded the sc_def_rate battle config. The new battle flags are mob_sc_def_rate, pc_sc_def_rate, mob_max_sc_def, pc_max_sc_def - Modified Marionette Control so that the max bonus stats you get is capped to your server's defined max stats instead of 99. - Mobs are no longer affected by the vs_traps_bctall switch. - Added function pc_damage_sp to damage the SP of players. - Modified Warm skills so that it only hurts SP of players while attacking/knocking back mobs. - Autospell loop breaks after one skill is successful.
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@5288 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/map/charcommand.c
@@ -1267,7 +1267,7 @@ int charcommand_baselevel(
 					level = pc_maxbaselv(pl_sd) - pl_sd->status.base_level;
 				for (i = 1; i <= level; i++)
 					pl_sd->status.status_point += (pl_sd->status.base_level + i + 14) / 5;
-				pl_sd->status.base_level += level;
+				pl_sd->status.base_level += (unsigned int)level;
 				clif_updatestatus(pl_sd, SP_BASELEVEL);
 				clif_updatestatus(pl_sd, SP_NEXTBASEEXP);
 				clif_updatestatus(pl_sd, SP_STATUSPOINT);
@@ -1290,7 +1290,7 @@ int charcommand_baselevel(
 						pl_sd->status.status_point = 0;
 					clif_updatestatus(pl_sd, SP_STATUSPOINT);
 				} // to add: remove status points from stats
-				pl_sd->status.base_level -= level;
+				pl_sd->status.base_level -= (unsigned int)level;
 				clif_updatestatus(pl_sd, SP_BASELEVEL);
 				clif_updatestatus(pl_sd, SP_NEXTBASEEXP);
 				status_calc_pc(pl_sd, 0);
@@ -1338,7 +1338,7 @@ int charcommand_joblevel(
 				if ((unsigned int)level > pc_maxjoblv(pl_sd) ||
 					pl_sd->status.job_level > pc_maxjoblv(pl_sd) -level)
 					level = pc_maxjoblv(pl_sd) - pl_sd->status.job_level;
-				pl_sd->status.job_level += level;
+				pl_sd->status.job_level += (unsigned int)level;
 				clif_updatestatus(pl_sd, SP_JOBLEVEL);
 				clif_updatestatus(pl_sd, SP_NEXTJOBEXP);
 				pl_sd->status.skill_point += level;
@@ -1354,7 +1354,7 @@ int charcommand_joblevel(
 				level*=-1;
 				if ((unsigned int)level >= pl_sd->status.job_level)
 					level = pl_sd->status.job_level-1;
-				pl_sd->status.job_level -= level;
+				pl_sd->status.job_level -= (unsigned int)level;
 				clif_updatestatus(pl_sd, SP_JOBLEVEL);
 				clif_updatestatus(pl_sd, SP_NEXTJOBEXP);
 				if (pl_sd->status.skill_point < level)
